Separate kafka readers to base package
